+++ generator/lib/src/generator.dart
@@ -742,12 +742,16 @@ class RetrofitGenerator extends GeneratorForAnnotation<retrofit.RestApi> {
       final queryParams = args.remove(_queryParamsVar)!;
       final baseUrl = args.remove(_baseUrlVar)!;
       final cancelToken = args.remove(_cancelToken);
+      final sendProgress = args.remove(_onSendProgress);
 
       final type = refer(_displayString(_getResponseType(m.returnType)));
 
       final composeArguments = <String,Expression>{_queryParamsVar: queryParams, _dataVar: dataVar};
       if (cancelToken != null) {
         composeArguments[_cancelToken] = cancelToken;
+      }
+      if (sendProgress != null){
+        composeArguments[_onSendProgress] = sendProgress;
       }
 
       return refer('_setStreamType').call([