+++ generator/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 2.0.0-beta3
 
-- Fixed CancelRequest generating invalid code 
+- Fixed CancelRequest generating invalid code (#322)
+- Fixed onSendProgress generating invalid code (#325)
 
 ## 2.0.0-beta2